+++ media/smart-card/examples/vr-flexible-ui-block-footer.tsx
@@ -4,10 +4,15 @@ import { jsx } from '@emotion/core';
 
 import { VRTestWrapper } from './utils/vr-test';
 import { FooterBlock } from '../src/view/FlexibleCard/components/blocks';
-import { ActionName, SmartLinkSize } from '../src/constants';
+import { SmartLinkSize } from '../src/constants';
 import { getCardState } from './utils/flexible-ui';
 import FlexibleCard from '../src/view/FlexibleCard';
 import { ActionItem } from '../src';
+import {
+  makeCustomActionItem,
+  makeDeleteActionItem,
+} from './flexible-ui/utils';
+import PremiumIcon from '@atlaskit/icon/glyph/premium';
 
 const renderFooter = (size?: SmartLinkSize, actions?: ActionItem[]) => {
   const cardState = getCardState();
@@ -18,52 +23,28 @@ const renderFooter = (size?: SmartLinkSize, actions?: ActionItem[]) => {
   );
 };
 
-const actions: ActionItem[] = [
-  {
-    name: ActionName.DeleteAction,
-    hideIcon: true,
-    onClick: () => console.log('delete'),
-  },
-];
+const actions: ActionItem[] = [makeDeleteActionItem()];
 
 export default () => (
   <VRTestWrapper title="Flexible UI: FooterBlock">
     <h5>Default</h5>
     {renderFooter()}
     <h5>With two actions</h5>
+    {renderFooter(
+      SmartLinkSize.Medium,
+      Array(2)
+        .fill(null)
+        .map(() => makeDeleteActionItem()),
+    )}
+    <h5>With 3 Custom actions</h5>
     {renderFooter(SmartLinkSize.Medium, [
-      {
-        name: ActionName.DeleteAction,
-        hideIcon: true,
-        onClick: () => console.log('delete'),
-      },
-      {
-        name: ActionName.DeleteAction,
-
-        hideIcon: true,
-        onClick: () => console.log('delete'),
-      },
-    ])}
-    <h5>With three actions</h5>
-    {renderFooter(SmartLinkSize.Medium, [
-      {
-        name: ActionName.DeleteAction,
-        testId: 'first-action-item',
-        hideIcon: true,
-        onClick: () => console.log('delete'),
-      },
-      {
-        name: ActionName.DeleteAction,
-        testId: 'second-action-item',
-        hideIcon: true,
-        onClick: () => console.log('delete'),
-      },
-      {
-        name: ActionName.DeleteAction,
+      makeCustomActionItem(),
+      makeDeleteActionItem(),
+      makeCustomActionItem({
+        icon: <PremiumIcon label="magic" />,
         testId: 'third-action-item',
-        hideIcon: true,
-        onClick: () => console.log('delete'),
-      },
+        content: 'Magic!',
+      }),
     ])}
     {Object.values(SmartLinkSize).map((size) => (
       <React.Fragment>