+++ media/smart-card/src/__tests__/visual-regression/flexible-card.test.tsx
@@ -15,7 +15,7 @@ describe('Flexible Card', () => {
       const url = getURL('vr-flexible-ui-block-title');
       const page = await setup(url);
       await page.waitForSelector('[data-testid="smart-links-container"]');
-      const image = await takeSnapshot(page, 1660);
+      const image = await takeSnapshot(page, 1760);
 
       expect(image).toMatchProdImageSnapshot();
     });