+++ media/smart-card/src/view/FlexibleCard/components/actions/index.ts
@@ -10,3 +10,6 @@ import { createUIAction } from './utils';
 export const DeleteAction = createUIAction<ActionProps>(
   ActionName.DeleteAction,
 );
+export const CustomAction = createUIAction<ActionProps>(
+  ActionName.CustomAction,
+);